🚑(cfg): Fix a missing property for the init scripts
This commit fixes the missing property by copying from the stock's
`/system/build.prop`.

The property `ro.product.model.num` is used in

 * `on_post-fs-data`, _/vendor/bin/init.fih.preConfig.sh_
 * `on_ffbm`, _/vendor/bin/init.qcom.sh_

It is **NOT** required to be put into `default.prop` but do so to obtain
higher priority than `/system/build.prop` in case some GSIs defined it.