✨(sys): Add SHARP stock boot animation
Here comes the stock boot animation that is copied from my 2.080WW.

    60a2054cd43a94c909033895acf6d85a06ad0bd8  system/CDA/300ID/Power-on-animation/bootanimation.zip

From the `cmds/bootanimation/BootAnimation.cpp` we can see the search
paths for **bootanimation.zip** are restricted to only `system/media`
and `oem/media`.

So we just maintain a valid directory structure with the bootanimation
files inside, and **mount** it when the system boots.